% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{theme_barnardos}
\alias{theme_barnardos}
\title{Add +theme_barnardos() to a ggplot2 object to apply default Barnardo's branding}
\usage{
theme_barnardos()
}
\description{
Add +theme_barnardos() to a ggplot2 object to apply default Barnardo's branding
}
